+++ tests/plugins/jobs/amqp/.rr-no-global.yaml
@@ -24,52 +24,24 @@ jobs:
 
   # list of broker pipelines associated with endpoints
   pipelines:
-    test-local:
-      driver: ephemeral
-      priority: 10
-      pipeline_size: 10000
-
-    test-local-2:
-      driver: ephemeral
-      priority: 1
-      pipeline_size: 10000
-
-    test-local-3:
-      driver: ephemeral
-      priority: 2
-      pipeline_size: 10000
-
     test-1:
       driver: amqp
       priority: 1
-      pipeline_size: 1000000
+      pipeline_size: 100
       queue: test-1-queue
       exchange: default
       exchange_type: direct
       routing_key: test
 
-    test-2-amqp:
+    test-2:
       driver: amqp
       priority: 2
-      pipeline_size: 100000
+      pipeline_size: 100
       queue: test-2-queue
       exchange: default
       exchange_type: direct
       routing_key: test-2
 
-    test-2:
-      driver: beanstalk
-      priority: 11
-      tube: default
-      pipeline_size: 1000000
-
-    test-3:
-      #      priority: 11 - not defined, 10 by default
-      # driver locality not specified, local by default
-      driver: sqs
-      pipeline_size: 1000000
-      queue: default
-
   # list of pipelines to be consumed by the server, keep empty if you want to start consuming manually
-  consume: [ "test-local", "test-local-2", "test-local-3", "test-1", "test-2-amqp" ]
+  consume: [ "test-1", "test-2" ]
