--- tests/plugins/jobs/configs/.rr-jobs-pause-resume-all.yaml
@@ -1,78 +1,0 @@
-rpc:
-  listen: tcp://127.0.0.1:6001
-
-server:
-  command: "php ../../client.php echo pipes"
-  relay: "pipes"
-  relay_timeout: "20s"
-
-amqp:
-  addr: amqp://guest:guest@localhost:5672/
-
-logs:
-  level: debug
-  mode: development
-
-jobs:
-  # num logical cores by default
-  num_pollers: 10
-  # 1mi by default
-  pipeline_size: 100000
-  # worker pool configuration
-  pool:
-    num_workers: 10
-    max_jobs: 0
-    allocate_timeout: 60s
-    destroy_timeout: 60s
-
-  # list of broker pipelines associated with endpoints
-  pipelines:
-    test-local:
-      driver: ephemeral
-      priority: 10
-      pipeline_size: 10000
-
-    test-local-2:
-      driver: ephemeral
-      priority: 1
-      pipeline_size: 10000
-
-    test-local-3:
-      driver: ephemeral
-      priority: 2
-      pipeline_size: 10000
-
-    test-1:
-      driver: amqp
-      priority: 1
-      pipeline_size: 1000000
-      queue: test-1-queue
-      exchange: default
-      exchange_type: direct
-      routing_key: test
-
-    test-2-amqp:
-      driver: amqp
-      priority: 2
-      pipeline_size: 100000
-      queue: test-2-queue
-      exchange: default
-      exchange_type: direct
-      routing_key: test-2
-
-    test-2:
-      driver: beanstalk
-      priority: 11
-      tube: default
-      pipeline_size: 1000000
-
-    test-3:
-      #      priority: 11 - not defined, 10 by default
-      # driver locality not specified, local by default
-      driver: sqs
-      pipeline_size: 1000000
-      queue: default
-
-  # list of pipelines to be consumed by the server, keep empty if you want to start consuming manually
-  consume: [ "test-local", "test-local-2", "test-local-3", "test-1", "test-2-amqp" ]
-
